Refine Homebrew setup in .zshenv
Update the Homebrew configuration in the `.zshenv` file to enhance
Homebrew initialization. Provide a default `HOMEBREW_PREFIX` that is set
to /opt/homebrew. Removed redundant PATH modification for Homebrew's bin
directory as it's handled by `brew shellenv`. Only set Homebrew env
variables when the `brew` executable is present and executable.