Update oh-my-posh configuration
Modify the `base.toml` configuration file to improve the display of the
Git segment and add a Python segment.

The Git segment template for the reference has been adjusted by adding a
leading space for better readability.

A new Python segment has been added to Python version and virtual
environment details. This segment uses the `nf-fa-python` icon
(`\uED1B`) to represent the Python environment and is configured with a
green foreground.